Enable services to start at boot time on redhat.
Replace service.running with cmd.run 'service restart' at the end of each sls. This ensures the service always restarts when the sls is run to load updated config.

Simpler and more reliable to restart services at the end of the sls instead of using service.running with watch statements that are complicated, platform dependent and often incomplete.

PNDA-2734