Remove EPEL from set of standard online repositories
The EPEL repository is not a part of Red Hat Enterprise Linux, although
we install some packages from this repository these are installed via
the PNDA mirror. A user would not necessarily expect to have this
repository included by default.

It is also seen to be by far the most unreliable repo, causing repeated
failures in provisioning.

PNDA-3215