Fix log overrun when applying the blueprint fails.
When applying the blueprint, the monitoring function reports the status
in a loop. The report logs over and over again the same failure details
which ends up clogging the log report.
This change avoid duplicating those big reports but merely logs that
fact that the same task is still in failure state.