PNDA-3600: Make the spark/MR cli wrapper the master system cli.
Use alternatives to prioritize the pnda spark cli wrapper (which
implements the pnda queue placement policy) over the default spark cli.
This allows frameworks that use the default system cli to benefit from the
pnda queue placement policy.
Also, refactored the placement policy to be more modular and introduced
a wild card for the mapper that matches every user of the system. This
wild card will be used initially to map every user to the default queue
and avoid introducing a dependency other issues that will introduce more
complex queue structures and associated ACL's.