Configure the default queue ACL the same for HDP as for CDH.
In CDH, the pnda user is only configured with submit clearance to the
default queue. HDP has now been brought in sync with this setup.
Although, in the current setup it doesn't make much difference as the
submit ACL are incremented with the admin ACL, it better reflect the
intention and would be cleaner should the submit and admin user not
coincide.