Refactor the yarn client cli wrappers to manipulate the PATH env
The previous implementation was manipulating the alternatives but this
doesn't appears to be robust against upgrades. During the upgrade the
alternatives are being manipulated by the hadoop distro itself hence
resetting the wrappers modification.
The new implementation hence now manipulates the PATH instead to insert
the wrapper in front of the wrapped cli managed by hadoop.

PNDA-4511